backend/requirements.txt
# =========================================================
# LEGALGRAM 2.0 BACKEND - Python Dependencies
# =========================================================
# Install: pip install -r requirements.txt
# =========================================================

# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0

# AI/LLM
groq==0.4.2

# CORS & Security
python-multipart==0.0.6
python-dotenv==1.0.0

# HTTP Client (for external calls if needed)
httpx==0.26.0

# Utilities
pydantic==2.5.3
pydantic-settings==2.1.0

# Development
python-jose[cryptography]==3.3.0

# =========================================================
# OPTIONAL: For Production
# =========================================================
# Redis for session storage
# redis==5.0.1
# aioredis==2.0.1

# Logging
# loguru==0.7.2

# Monitoring
# prometheus-client==0.19.0
